tools/requirements.txt
mysql-connector
elasticsearch
ftfy